new function: consult-emms--get-artist-tracks
Abstract out from consult-emms--add-artist. This will be useful when
writing functions for interacting with artists, artists and
tracks. (e.g. in embark actions)

(this is a VERY simple function, mostly defined so that albums and
arists can be handled with very similar code, to keep the patterns the
same as much as possible.)